phong shading

+++ src/commons/texture.cpp
@@ -63,8 +63,8 @@ void Texture::generateProcedural(int width, int height) {
 
     SimplexNoise noise = SimplexNoise();
 
-    for (float x = 0; x < width; x++) {
-        for (float y = 0; y < height; y++) {
+    for (int x = 0; x < width; x++) {
+        for (int y = 0; y < height; y++) {
             int i = (y * width + x) * 4;
 
             //MARBLE
